Add files via upload
Credit card fraud is a common problem that costs millions of dollars to financial institutions and cardholders. Fraudsters use various techniques to deceive banks and individuals, making it challenging to detect fraudulent transactions. Therefore, there is a need to develop a reliable system that can identify fraudulent transactions and minimize the losses incurred.

In this project, we used supervised learning to attempt to classify certain test data as fraudulent or non-fraudulent through the training of our model. For this purpose, we will utilize the logistic regression model as helps to calculate the possibility of a particular event taking place (i.e., a transaction being fraudulent or not). We will also evaluate the performance of the Logistic Regression Model by comparing its performance with that of other Machine Learning approaches 

The dataset used for demonstration is: https://www.kaggle.com/datasets/mlg-ulb/creditcardfraud

The project utilizes Jupyter Notebook. Also, the following dependencies are used to implement the project: Pandas, Numpy, Tensorflow, Tabulate and Sci-kit (Sklearn).